sn-gamestate/pyproject.toml
[tool.poetry]
name = "sn-gamestate"
version = "0.2.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "sn_gamestate"}]

[tool.poetry.dependencies]
python = "^3.9"
tracklab = {path = "../tracklab", develop = true}
easyocr = "^1.7.1"
soccernet = "^0.1.55"
torch = "1.13.1"
torchvision = "0.14.1"
tracklab-calibration = {path = "plugins/calibration", develop = true}
prtreid = {git = "https://github.com/VlSomers/prtreid"}
mmocr = "^1.0.1"
mmpose = "^1.2.0"
open_clip_torch="*"
ultralytics = "8.0.100"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.tracklab_plugin]
soccernet = "sn_gamestate.config_finder:ConfigFinder"
